// cargodelivery-front/generated/models/RoadIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Point } from './Point';
import {
    PointFromJSON,
    PointFromJSONTyped,
    PointToJSON,
} from './Point';

/**
 * 
 * @export
 * @interface RoadIn
 */
export interface RoadIn {
    /**
     * 
     * @type {string}
     * @memberof RoadIn
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof RoadIn
     */
    description: string;
    /**
     * 
     * @type {Array<Point>}
     * @memberof RoadIn
     */
    path: Array<Point>;
}

/**
 * Check if a given object implements the RoadIn interface.
 */
export function instanceOfRoadIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "description" in value;
    isInstance = isInstance && "path" in value;

    return isInstance;
}

export function RoadInFromJSON(json: any): RoadIn {
    return RoadInFromJSONTyped(json, false);
}

export function RoadInFromJSONTyped(json: any, ignoreDiscriminator: boolean): RoadIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'],
        'path': ((json['path'] as Array<any>).map(PointFromJSON)),
    };
}

export function RoadInToJSON(value?: RoadIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'description': value.description,
        'path': ((value.path as Array<any>).map(PointToJSON)),
    };
}