// cargodelivery-front/generated/models/Point.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Point
 */
export interface Point {
    /**
     * longitude for coordinate
     * @type {number}
     * @memberof Point
     */
    longitude: number;
    /**
     * latitude for coordinate
     * @type {number}
     * @memberof Point
     */
    latitude: number;
}

/**
 * Check if a given object implements the Point interface.
 */
export function instanceOfPoint(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "longitude" in value;
    isInstance = isInstance && "latitude" in value;

    return isInstance;
}

export function PointFromJSON(json: any): Point {
    return PointFromJSONTyped(json, false);
}

export function PointFromJSONTyped(json: any, ignoreDiscriminator: boolean): Point {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'longitude': json['longitude'],
        'latitude': json['latitude'],
    };
}

export function PointToJSON(value?: Point | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'longitude': value.longitude,
        'latitude': value.latitude,
    };
}